client: post command
Exception handling brakes the logic of lense some reading were required
for good wrapper function (safePost).

The (^.) was a big help but I made some sideways until I read the the
good documentation (they brake the usage of the oprators between
versions).

Haskell has some addiction towards special operators which was a help in
case of ReadP but lense tend to overuse them.